Rename skill command server helper
