+++ packages/web/src/app/(app)/chat/[id]/page.tsx
@@ -19,7 +19,7 @@ import { env } from '@sourcebot/shared';
 import { hasEntitlement } from '@/lib/entitlements';
 import { ChatEntitlementMessage } from '@/features/chat/components/chatEntitlementMessage';
 import { captureEvent } from '@/lib/posthog';
-import { listAgentSkillCommandsOrEmpty } from '@/ee/features/chat/skills/commands.server';
+import { listAgentSkillCommandsOrEmpty } from '@/ee/features/chat/skills/skillCommands.server';
 
 interface PageProps {
     params: Promise<{
